swap tuna to new thing
